requirements.txt
adbutils==0.13.1
airtest==1.2.4
apkutils2==1.0.0
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.10
cigam==0.0.3
decorator==5.1.1
Deprecated==1.2.13
deprecation==2.1.0
facebook-wda==1.4.6
filelock==3.4.2
idna==3.3
Jinja2==3.0.3
logzero==1.7.0
lxml==4.7.1
MarkupSafe==2.0.1
mss==6.1.0
numpy==1.19.3
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
packaging==20.9
Pillow==9.0.0
progress==1.6
py==1.11.0
pyelftools==0.27
pyparsing==3.0.7
python-xlib==0.31
pywinauto==0.6.3
requests==2.27.1
retry==0.9.2
six==1.16.0
uiautomator2==2.16.13
urllib3==1.26.8
websocket-client==1.2.3
websockets==10.1
whichcraft==0.6.1
wrapt==1.13.3
xmltodict==0.12.0
